+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- Make `width` and `height` optional for the `.sized` extension.
+- Remove `child` parameter from the `notificationListener` extension.
+- Add extensions for `Transform.rotate`, `.scale`, `.translate`, and `.flip`.
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+Initial release